+++ src/yafowil/utils.py
@@ -171,6 +171,22 @@ def attr_value(key, widget, data, default=None):
     return attr
 
 
+def generic_html5_attrs(data_dict):
+    data_attrs = {}
+    for key, val in data_dict.items():
+        if val is None:
+            continue
+        ret = json.dumps(val) # js-ify
+        if isinstance(val, basestring):
+            ret = ret.strip('"') # for strings, remove leading and trailing
+                                 # double quote, since they are not needed for
+                                 # data-attributes
+        # replace camelCase with camel-case
+        key = re.sub("([a-z])([A-Z])","\g<1>-\g<2>", key).lower()
+        data_attrs['data-%s' % key] = ret
+    return data_attrs
+
+
 def data_attrs_helper(widget, data, attrs):
     """Creates a dictionary of JSON encoded data-attributes from a list of
     attribute-keys, ready to inject to a tag-renderer as expanded keyword
@@ -212,7 +228,6 @@ def data_attrs_helper(widget, data, attrs):
                                  # double quote, since they are not needed for
                                  # data-attributes
         # replace camelCase with camel-case
-        # XXX: why? -rnix
         key = re.sub("([a-z])([A-Z])","\g<1>-\g<2>", key).lower()
         data_attrs['data-%s' % key] = ret
     return data_attrs